requirements.txt
telethon
pyjwt
python-dotenv
requests
pytz
